Implement unified async road generation with deterministic IDs and lifecycle management
- Integrated roads into async chunking system for smooth 60+ FPS
- Changed road IDs from u32 to u64 with zigzag encoding for global uniqueness
- Implemented generate_unique_road_id(cell_coord, local_index) for collision-free IDs
- Added per-cell deterministic RNG (seeded from cell coordinates)
- Created RoadOwnership resource to track road→(chunk, entity) mappings
- Added proper cleanup: roads removed from RoadNetwork when chunks unload
- Roads tracked separately via RoadOwnership (not in chunk.entities)
- RoadBlueprint generation happens async off main thread
- Main thread spawns entities with proximity-based detail (<100m for markings)
- Deprecated legacy road_network_system (replaced by async pipeline)

Benefits:
- No duplicate roads (deduplication via road_id HashMap check)
- No memory leaks (roads evicted when owner chunk unloads)
- Deterministic: same cell always generates identical roads
- Thread-safe: async blueprint generation, main-thread-only spawning
- No ID collisions: zigzag encoding handles negative coordinates correctly

Test results: 187 roads stable, no panics, smooth streaming

Amp-Thread-ID: https://ampcode.com/threads/T-e08ed1e5-03a6-4b1d-8a11-4ca592dd25e4

+++ src/components/world.rs
@@ -232,7 +232,7 @@ pub fn cullable_to_visibility_range(
 // Road system components
 #[derive(Component)]
 pub struct RoadEntity {
-    pub road_id: u32,
+    pub road_id: u64,
 }
 
 #[derive(Component)]
